chore: optimize runtime code import
