+++ docs/.island/styles/index.css
@@ -1,6 +1,7 @@
-:root {
+/* Test for custom global style */
+/* :root {
   --island-c-brand: #5c6ac4;
   --island-c-brand-light: #7c87cf;
   --island-c-brand-dark: #4b5abf;
   --island-c-brand-darker: #3a49ba;
-}
+} */
