+++ src/shared/types/type.d.ts
